Feature: check for updated binaries
A file on a remote server (version.txt for example) that contains
the newest wallet information is retrieved, then the string in
that file is parsed and compared with the internal wallet version,
which is defined in src/version.h. If any integer in the remote
version is larger than the corresponding internal integer, a
clickable label is displayed in the GUI overview page, prompting
the user to download the new wallet version. This check occurs
once per day, as commited.